Convert windows strange init requirements to just use our initfuncs
