+++ src/initfuncs.c
@@ -5,13 +5,14 @@
  * Wrapper to call initfuncs
  */
 
-// prototype any internal (non-public) initfuncs
+// prototype any internal (non-public) initfuncs (always sorted!)
 void n3n_initfuncs_conffile_defs ();
 void n3n_initfuncs_metrics ();
 void n3n_initfuncs_pearson ();
 void n3n_initfuncs_peer_info ();
 void n3n_initfuncs_random ();
 void n3n_initfuncs_transform ();
+void n3n_initfuncs_win32 ();
 
 void n3n_initfuncs () {
     // TODO:
@@ -19,6 +20,12 @@ void n3n_initfuncs () {
     //   constructors within their own object file, but we need to reference
     //   them externally or the linker will never link that object
 
+    // (must be first because windows is crazy)
+#ifdef _WIN32
+    n3n_initfuncs_win32();
+#endif
+
+    // (sorted list)
     n3n_initfuncs_conffile_defs();
     n3n_initfuncs_metrics();
     n3n_initfuncs_pearson();
